# docs/manual-payments.openapi.yml
openapi: 3.0.3
info:
  title: Anytimebot — Pagos manuales y ventas no agendadas
  version: 1.0.0
  description: |
    Módulo de ventas y cobros manuales. Todos los importes van en **céntimos
    enteros** (`amountCents`) y la moneda es la del tenant en minúsculas.
    Los pagos nunca se borran: se anulan (con motivo) o se reembolsan.
    Stripe solo interviene en pagos online (`CARD_ONLINE` con
    `stripePaymentIntent` real); los métodos manuales jamás llaman a Stripe.
    Autenticación: cookie de sesión NextAuth (dashboard).
servers:
  - url: https://anytimebot.app
    description: Producción
  - url: http://localhost:3000
    description: Desarrollo
...
components:
...
  securitySchemes:
    nextAuthCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
security:
  - nextAuthCookie: []
paths:
...
  /api/reports/payments:
    get:
      tags: [reports]
      summary: Reporte de pagos (bruto, reembolsos, neto) agrupado
      parameters:
        - { name: groupBy, in: query, schema: { type: string, enum: [method, staff, location, day], default: method } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { name: format, in: query, schema: { type: string, enum: [json, csv], default: json } }
      responses:
        '200':
          description: Filas agregadas + totales (o CSV con BOM y ; como separador)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      groupBy: { type: string }
                      rows:
                        type: array
                        items:
                          type: object
                          properties:
                            key: { type: string }
                            label: { type: string }
                            grossCents: { type: integer }
                            refundCents: { type: integer }
                            netCents: { type: integer }
                            count: { type: integer }
                      totals:
                        type: object
                        properties:
                          grossCents: { type: integer }
                          refundCents: { type: integer }
                          netCents: { type: integer }
                          count: { type: integer }
                      currency: { type: string }
            text/csv:
              schema: { type: string }